Set Close button `type` attribute rather than property
IE8 does not allow you to change the `type` property of a button after
it has been created (eg: by createElement). But it *does* seem to allow
you to change the `type` *attribute*, which is good enough for our needs
(preventing the parent form from being submitted when the Close button
on a Taggle tag is clicked).